+++ recurtools/__init__.py
@@ -1,13 +1,8 @@
-# noqa: D104
-
 from .nested import nested
 from .utils import (
     countrecursive,
     flatten,
     indexrecursive,
-    inrecursive,
     starchain,
     sumrecursive,
 )
-
-__all__ = ["flatten", "lenrecursive", "sumrecursive", "countrecursive", "inrecursive", "indexrecursive", "starchain", "nested"]  # noqa: E501