+++ recurtools/nested.py
@@ -1,7 +1,8 @@
 # noqa: D100
-from typing import Collection, Container
+from contextlib import suppress
+from typing import Any, Collection, Container, Iterable
 
-from recurtools.utils import countrecursive, flatten, inrecursive
+from recurtools.utils import countrecursive, flatten
 
 
 class nested(Collection):  # noqa: N801
@@ -43,13 +44,23 @@ class nested(Collection):  # noqa: N801
     def __init__(self, nestedcontainer: Container) -> None:
         self.nestedcontainer = nestedcontainer
 
-    def __contains__(self, __o: object) -> bool:  # noqa: D105
-        return inrecursive(self.nestedcontainer, __o)
 
-    def __len__(self):  # noqa: ANN204, D105
+    def __contains__(self, __other: Any) -> bool:
+        
+        def _in(collection: Iterable, val: Any):
+            for item in collection: # by using flatten in the call to _in, we are guaranteed an initial iterable
+                if item == val: return True
+                with suppress(TypeError): # item is not guaranteed to support __contains__
+                    if val in item: return True
+            return False
+
+        return _in(flatten(self.nestedcontainer), __other)
+
+
+    def __len__(self):  # noqa: ANN204
         return len(list(flatten(self.nestedcontainer)))
 
-    def __iter__(self):  # noqa: ANN204, D105
+    def __iter__(self):  # noqa: ANN204
         return flatten(self.nestedcontainer)
 
     def count(self, __o):  # noqa: ANN001, ANN201, D102